+++ Libraries/LibIPC/ClientConnection.h
@@ -108,7 +108,7 @@ public:
 
         auto buffer = message.encode();
 
-        int nwritten = write(m_socket->fd(), buffer.data(), (size_t)buffer.size());
+        int nwritten = write(m_socket->fd(), buffer.data(), buffer.size());
         if (nwritten < 0) {
             switch (errno) {
             case EPIPE:
@@ -149,7 +149,7 @@ public:
         }
 
         size_t decoded_bytes = 0;
-        for (size_t index = 0; index < (size_t)bytes.size(); index += decoded_bytes) {
+        for (size_t index = 0; index < bytes.size(); index += decoded_bytes) {
             auto remaining_bytes = ByteBuffer::wrap(bytes.data() + index, bytes.size() - index);
             auto message = Endpoint::decode_message(remaining_bytes, decoded_bytes);
             if (!message) {
